requirements.txt
fastapi~=0.111.0
starlette~=0.37.2
python-dotenv~=1.0.1
pydantic~=2.8.2
uvicorn~=0.30.1
mysqlclient~=2.2.4
SQLAlchemy~=2.0.31
bcrypt~=4.1.3
python-multipart~=0.0.9